% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.R
\name{as.data.frame.trios}
\alias{as.data.frame.trios}
\title{Convert a trios object to a data frame in long format}
\usage{
\method{as.data.frame}{trios}(trios)
}
\arguments{
\item{trios}{}
}
\value{
A data frame in long form with columns for depth, wavelength, sensor type and irradiance value
}
\description{
Convert a trios object to a data frame in long format
}
