% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pool_vol.R
\name{calc_pool_vol}
\alias{calc_pool_vol}
\title{Get volume for pooling in equal concentrations}
\usage{
calc_pool_vol(plates, max_vol, blank_threshold, blank_vol = 10)
}
\arguments{
\item{plates}{A data frame in long format, with concentrations for each well named conc_ng_ul,
typically the output of fluorescence_to_conc. Can contain multiple plates in a single data frame.}

\item{max_vol}{The maximum volume to use from a single sample. For example the total volume of your sample - 5.}

\item{blank_threshold}{Maximum concentration for something to be considered a "blank", which will be pooled with a fixed volume}

\item{blank_vol}{Fixed volume to use for samples with concentrations below blank_threshold}
}
\value{
The input data frame with an added column pool_vol
}
\description{
Get volume for pooling in equal concentrations
}
