% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluorescence_to_conc.R
\name{fluorescence_to_conc}
\alias{fluorescence_to_conc}
\title{Convert fluorescence values from a plate to concentrations of your target substance based on a series of standards.}
\usage{
fluorescence_to_conc(plate, std, sample_volume, well_volume = 200)
}
\arguments{
\item{plate}{A data.frame formatted as a well plate, with names A:H in the first column,
and the subsequent 12 columns containing fluorescence reads. The rest of the columns will be ignored. This is compatible with a single
plate from the plate reader at AQUA read from an excel file.}

\item{std}{Either a data frames with two columns named fluorescence and conc_ng_ul, or a
named vector containing concentrations and the names of the wells containing the respective concentrations.}

\item{sample_volume}{The volume of sample added to each well}

\item{well_volume}{The total volume of the well}
}
\value{
A data frame in long format with concentrations for each well
}
\description{
Made specifically for quantifying DNA with picogreen, but can probably
also be used for other plate reader applications
}
