% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_to_surface.R
\name{normalize_to_surface}
\alias{normalize_to_surface}
\title{Normalize down- and upwelling irradiance to surface irradiance}
\usage{
normalize_to_surface(trios, relative_to_surface = TRUE)
}
\arguments{
\item{trios}{A trios object}
}
\value{
A normalized trios object
}
\description{
Normalize down- and upwelling irradiance to surface irradiance
}
