% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.trios}
\alias{plot.trios}
\title{Make a summary plot of a trios object}
\usage{
\method{plot}{trios}(trios, title = NULL)
}
\arguments{
\item{trios}{A trios object}
}
\description{
Make a summary plot of a trios object
}
\examples{

plot(trios_march)
plot(trios_september)

}
