% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_plate.R
\name{plot_plate}
\alias{plot_plate}
\title{Plot well-plates with a fill-value for each well}
\usage{
plot_plate(plate, fill_var = conc_ng_ul, round_dec = 2)
}
\arguments{
\item{fill_var}{The variable to use for fill and labeling, e.g. conc_ng_ul or pool_vol}

\item{round_dec}{Number of decimals to use for rounding of the labels}

\item{plates}{A data frame in long format, with wells and either concentrations or pool volumes,
typically the output of fluorescence_to_conc or calc_pool_vol. If the data frame contains a column
"plate", the resulting plot will be faceted by plate.}
}
\value{
A ggplot object, plotting the plate with its physical layout,
and geom_tile and geom_text for the variable of interest.
}
\description{
Plot well-plates with a fill-value for each well
}
