% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_trios.R
\name{read_trios}
\alias{read_trios}
\title{Title}
\usage{
read_trios(
  path,
  depths,
  sensor_air = "8175",
  sensor_up = "8078",
  sensor_down = "501A",
  sensor_down_alt = "817c"
)
}
\arguments{
\item{path}{}

\item{depths}{}

\item{sensor_air}{}

\item{sensor_up}{}

\item{sensor_down}{}

\item{sensor_down_alt}{}
}
\value{

}
\description{
Title
}
