% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.trios.R
\name{subset.trios}
\alias{subset.trios}
\title{Subset a trios object to wavelengths and depths}
\usage{
\method{subset}{trios}(trios, lower = NULL, upper = NULL, depths = NULL)
}
\arguments{
\item{trios}{A trios object}

\item{lower}{Lower wavelength (non-inclusive)}

\item{upper}{Upper wavelength (non-inclusive)}

\item{depths}{the depths to subset to}
}
\value{
a trios object containing the specified wavelengths and depths
}
\description{
Subset a trios object to wavelengths and depths
}
