% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.trios.R
\name{subset_depth}
\alias{subset_depth}
\title{Subset a trios object to certain depths}
\usage{
subset_depth(trios, depths)
}
\arguments{
\item{trios}{A trios object}

\item{depths}{the depths to subset to}
}
\value{
A trios object containing only the specified depths
}
\description{
Subset a trios object to certain depths
}
