% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_to_par.R
\name{subset_to_wavelength}
\alias{subset_to_wavelength}
\title{Subset a trios object to certain wavelengths}
\usage{
subset_to_wavelength(trios, lower = 400, upper = 700)
}
\arguments{
\item{trios}{A trios object}

\item{lower}{Lower wavelength (non-inclusive)}

\item{upper}{Upper wavelength (non-inclusive)}
}
\value{
A trios object containing only values between the specified wavelengths
}
\description{
Subset a trios object to certain wavelengths
}
