% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trios_march.R
\docType{data}
\name{trios_march}
\alias{trios_march}
\title{Trios data from the Oslofjord in March 2016}
\format{
\subsection{\code{trios_march}}{

A trios object with 10 depths and 255 wavelengths
}
}
\usage{
trios_march
}
\description{
A data set with 10 depths from 0-9 meters from the Oslofjord in March 2016
}
\keyword{datasets}
