% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trios_september.R
\docType{data}
\name{trios_september}
\alias{trios_september}
\title{Trios data from the Oslofjord in September 2022}
\format{
\subsection{\code{trios_september}}{

A trios object with 2 depths and 255 wavelengths
}
}
\usage{
trios_september
}
\description{
A data set with 2 depths, 1 and 5 meters, from the Oslofjord in September 2022
}
\keyword{datasets}
