% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trios_sync_sensors.R
\name{trios_sync_sensors}
\alias{trios_sync_sensors}
\title{Title}
\usage{
trios_sync_sensors(
  trios_long,
  depths,
  sensor_air = "8175",
  sensor_up = "8078",
  sensor_down = "501A",
  sensor_down_alt = "817c"
)
}
\arguments{
\item{sensor_air}{}

\item{sensor_up}{}

\item{sensor_down}{}
}
\value{

}
\description{
Title
}
