% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trios_to_df.R
\name{trios_to_df}
\alias{trios_to_df}
\title{Convert a trios object to a data frame in long format}
\usage{
trios_to_df(trios)
}
\arguments{
\item{trios}{}
}
\value{
A data frame in long form with columns for depth, wavelength, sensor type and irradiance value
}
\description{
Convert a trios object to a data frame in long format
}
