% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_trios.R
\name{read_trios}
\alias{read_trios}
\title{Read a set of Trios files}
\usage{
read_trios(
  path,
  depths,
  sensor_air = "8175",
  sensor_up = "8078",
  sensor_down = "501A",
  sensor_down_alt = "817c"
)
}
\arguments{
\item{path}{Path to parent folder of the Trios data}

\item{depths}{Sample depths for Trios measurements. This needs to have the same number of elements as there are files per sensor.}

\item{sensor_air}{Name of the air sensor}

\item{sensor_up}{Name of the upwards facing sensor}

\item{sensor_down}{Name of the downward facing sensor}

\item{sensor_down_alt}{Alternative name of downward facing sensor, in case there are multiple}
}
\value{
A trios object, containing sensors "air", "up" and "down", and wavelengths "w" and depth "depth"
}
\description{
Read a set of Trios files
}
