% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trios_sync_sensors.R
\name{trios_sync_sensors}
\alias{trios_sync_sensors}
\title{Sync the wavelengths of Trios sensors and make into Trios object}
\usage{
trios_sync_sensors(
  trios_long,
  depths,
  sensor_air = "8175",
  sensor_up = "8078",
  sensor_down = "501A",
  sensor_down_alt = "817c"
)
}
\arguments{
\item{depths}{Sample depths for Trios measurements. This needs to have the same number of elements as there are files per sensor.}

\item{sensor_air}{Name of the air sensor}

\item{sensor_up}{Name of the upwards facing sensor}

\item{sensor_down}{Name of the downward facing sensor}

\item{sensor_down_alt}{Alternative name of downward facing sensor, in case there are multiple}
}
\value{
A trios object, containing sensors "air", "up" and "down", and wavelengths "w" and depth "depth"
}
\description{
Sync the wavelengths of Trios sensors and make into Trios object
}
